refactor(pipeline): PipelineKernel を RecordingOrchestrator に統合
PBI 03 (architecture deepening 2026-09-03b, RICE 6.0):

- PipelineKernel(60行の thin loop)を RecordingOrchestrator.executeInternal
  に inline 化し削除 — sole state owner 化
- pipeline セマンティクス(PrivatePage/Duplicate 特例、FATAL/RETRY
  error mapping、previewBreakpoint)が step 定義と同一箇所に集約
- deletion test: Kernel 削除で複雑さは Orchestrator 内に集中
  （外部 caller はゼロだったため純粋な interface 削減）

既存テスト: RecordingPipeline* 4ファイル 44 tests + coverage 2ファイル
45 tests 全 green
